update the quantize script

+++ gemma4/quantize.py
@@ -171,7 +171,7 @@ def prepare_int4_weight_and_scales_and_zeros(weight_bf16, groupsize, inner_k_til
         weight_bf16, n_bit=4, groupsize=groupsize
     )
     weight_int4pack = torch.ops.aten._convert_weight_to_int4pack(
-        weight_int32, inner_k_tiles
+        weight_int32.to(torch.uint8), inner_k_tiles
     )
     return weight_int4pack, scales_and_zeros
